requirements.txt
 numpy
scipy
pandas
loguru
flash-attn
PyWavelets
matplotlib
tqdm
sktime
PyYAML
json-repair
timm